fix: move shared code into new library
