+++ test/spec/modules/vidazooBidAdapter_spec.js
@@ -1,7 +1,12 @@
 import {expect} from 'chai';
 import {
   spec as adapter,
+  storage,
   createDomain,
+  getVidazooSessionId,
+  webSessionId
+} from 'modules/vidazooBidAdapter.js';
+import {
   hashCode,
   extractPID,
   extractCID,
@@ -11,9 +16,8 @@ import {
   tryParseJSON,
   getUniqueDealId,
   getNextDealId,
-  getVidazooSessionId,
-  webSessionId
-} from 'modules/vidazooBidAdapter.js';
+  getTopWindowQueryParams
+} from 'libraries/vidazooUtils/bidderUtils.js'
 import * as utils from 'src/utils.js';
 import {version} from 'package.json';
 import {useFakeTimers} from 'sinon';
@@ -204,15 +208,6 @@ const REQUEST = {
     bidId: '2d52001cabd527'
   }
 };
-
-function getTopWindowQueryParams() {
-  try {
-    const parsedUrl = utils.parseUrl(window.top.document.URL, {decodeSearchAsString: true});
-    return parsedUrl.search;
-  } catch (e) {
-    return '';
-  }
-}
 
 describe('VidazooBidAdapter', function () {
   describe('validtae spec', function () {
@@ -829,15 +824,15 @@ describe('VidazooBidAdapter', function () {
     const key = 'myDealKey';
 
     it('should get the next deal id', function () {
-      const dealId = getNextDealId(key);
-      const nextDealId = getNextDealId(key);
+      const dealId = getNextDealId(storage, key);
+      const nextDealId = getNextDealId(storage, key);
       expect(dealId).to.be.equal(1);
       expect(nextDealId).to.be.equal(2);
     });
 
     it('should get the first deal id on expiration', function (done) {
       setTimeout(function () {
-        const dealId = getNextDealId(key, 100);
+        const dealId = getNextDealId(storage, key, 100);
         expect(dealId).to.be.equal(1);
         done();
       }, 200);
@@ -858,13 +853,13 @@ describe('VidazooBidAdapter', function () {
     const key = 'myKey';
     let uniqueDealId;
     beforeEach(() => {
-      uniqueDealId = getUniqueDealId(key, 0);
+      uniqueDealId = getUniqueDealId(storage, key, 0);
     })
 
     it('should get current unique deal id', function (done) {
       // waiting some time so `now` will become past
       setTimeout(() => {
-        const current = getUniqueDealId(key);
+        const current = getUniqueDealId(storage, key);
         expect(current).to.be.equal(uniqueDealId);
         done();
       }, 200);
@@ -872,7 +867,7 @@ describe('VidazooBidAdapter', function () {
 
     it('should get new unique deal id on expiration', function (done) {
       setTimeout(() => {
-        const current = getUniqueDealId(key, 100);
+        const current = getUniqueDealId(storage, key, 100);
         expect(current).to.not.be.equal(uniqueDealId);
         done();
       }, 200)
@@ -896,8 +891,8 @@ describe('VidazooBidAdapter', function () {
         shouldAdvanceTime: true,
         now
       });
-      setStorageItem('myKey', 2020);
-      const {value, created} = getStorageItem('myKey');
+      setStorageItem(storage, 'myKey', 2020);
+      const {value, created} = getStorageItem(storage, 'myKey');
       expect(created).to.be.equal(now);
       expect(value).to.be.equal(2020);
       expect(typeof value).to.be.equal('number');
@@ -908,7 +903,7 @@ describe('VidazooBidAdapter', function () {
     it('should get external stored value', function () {
       const value = 'superman'
       window.localStorage.setItem('myExternalKey', value);
-      const item = getStorageItem('myExternalKey');
+      const item = getStorageItem(storage, 'myExternalKey');
       expect(item).to.be.equal(value);
     });
